updated babel setup and moved index.js into src
