+++ .babelrc
@@ -1,11 +1,4 @@
 {
-  "presets": [
-    "es2015",
-    "react",
-    "stage-2"
-  ],
-  "plugins": [
-    "transform-runtime",
-    "transform-async-to-generator"
-  ]
+    "presets": ["@babel/preset-env", "@babel/preset-react"],
+    "plugins": ["@babel/plugin-transform-runtime"]
 }
