--- index.js
@@ -1,3 +1,0 @@
-global.regeneratorRuntime = require('babel-regenerator-runtime');
-require("babel-register")();
-require("./src/index.js");
