+++ package.json
@@ -1,11 +1,12 @@
+
 {
   "name": "resonators-server",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": ". ~/.nvm/nvm.sh && nvm use && ENV=dev node index.js",
-    "win-start": "nvm use && SET ENV=dev && node index.js",
+    "start": ". ~/.nvm/nvm.sh && nvm use && ENV=dev node src/index.js",
+    "win-start": "nvm use && SET ENV=dev && node src/index.js",
     "test": ". ~/.nvm/nvm.sh && nvm use && ./test/run -- --watch test/**/*.test.js",
     "win-test": "nvm use && ./test/run -- --watch test/**/*.test.js",
     "test:noinit": ". ~/.nvm/nvm.sh && nvm use && ./test/run noinit -- --watch test/**/*.test.js",
@@ -17,12 +18,13 @@
     "initdb": "./scripts/initdb",
     "win-initdb": "node scripts\\initdb",
     "storybook": "start-storybook -p 9007 -c .storybook",
-    "start:prod": "sudo NODE_ENV=production ENV=production pm2 -f start index.js --name resonators",
+    "start:prod": "sudo NODE_ENV=production ENV=production pm2 -f start src/index.js --name resonators",
     "restart:prod": "sudo pm2 restart resonators"
   },
   "author": "InterAl",
   "license": "ISC",
   "dependencies": {
+    "@babel/runtime": "^7.11.2",
     "aws-sdk": "^2.751.0",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
@@ -48,10 +50,13 @@
     "winston": "^3.3.3"
   },
   "devDependencies": {
+    "@babel/core": "^7.11.6",
+    "@babel/node": "^7.10.5",
+    "@babel/plugin-transform-runtime": "^7.11.5",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "@babel/register": "^7.11.5",
     "babel-eslint": "^10.1.0",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-regenerator": "^6.26.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
     "chai": "^4.2.0",
     "chai-shallow-deep-equal": "^1.4.6",
     "eslint": "^7.8.1",
