+++ src/index.js
@@ -1,22 +1,2 @@
-import api from "./api";
-import cfg from "./cfg";
-import initInfra from "./infra";
-import emailSchedulingJob from "./emailScheduler";
-import { setVapidKeys } from "./emailScheduler/push";
-// import syncCalendarsJob from "./calendars/calendarsSync";
-
-function startServer() {
-    api.listen(cfg.port, "0.0.0.0", () => {
-        console.log(`API running on port ${cfg.port} with ${process.env.ENV.toUpperCase()} environment`);
-    });
-}
-
-function startJobs() {
-    if (cfg.emailSchedulerOn) emailSchedulingJob.start();
-    // syncCalendarsJob.start();
-}
-
-initInfra();
-setVapidKeys();
-startJobs();
-startServer();
+require("@babel/register");
+require("./server");
